Gauss Seidel Power Flow Analysis
Matlab code for Gauss Seidel Power Flow Analysis for IEEE 300 Bus data. It also gives a graph of the convergence showing the no. of iterations it takes the tolerance to go below a limit. It shows that the lower we keep the tolerance the more iterations it takes for the graph to converge.